Fix major data leak vulnerability
There was a major bug in the function 'noAccessIfNotLoggedIn' which lead to data leak when a malicious user simply clicked view source that user was able to see database details that should not visible to outside unauthorised users. The problem code was using javascript to redirect a user to the index page when not authorized however javascript 'window.location' is not sufficient to redirect the browser's view-source option and hence data was leaked.